Token parsing modified to use a ReadableTrackingBuffer, removing the need to track the position within the buffer within the token parsing.
This is a complex and wide ranging change, that significantly simplifies token parsing, and makes it easier to read and maintain.
Any attempt to read past the end of the available buffer results in an exception being thrown. This may have a small runtime overhead, but it's worth it because of the simplified code.

Functions added to ReadableTrackingBuffer for more data types.
Functions added to WritableTrackingBuffer for more data types.
Many unit tests converted to use WritableTrackingBuffer, to avoid them needing to track the position withing the buffer.