Use iconv-lite instead of iconv.
The iconv package is not supported on Windows. This meant that it was included as an
optional dependency, and codepages were not properly supported on Windows.
The iconv package also caused confusion with Windows users, as the installation error
messages looked alarming despite npm recovering the installation.

Using iconv lite takes tedious back to being pure Javascript for itself and its
dependencies.