Clarify that options takes a database parameter
Hi,

First, many thanks for putting this together. 

I just got (stupidly) stuck on this for a bit until I read the docs a bit more closely and realized that the DB name is part of the options object. I figured I would add a couple of lines to this to clarify even further for developers down the road. Many thanks!