Fix VarBinary handling with strings.
When storing a String into a VarBinary column, tedious currently
incorrectly encodes that string as UCS-2. This is a bit weird, because
that means that

```
request.addParameter('test', TYPES.VarBinary, "Some String")
```

and

```
request.addParameter('test', TYPES.VarBinary, new Buffer("Some String"))
```

will cause different binary data to be stored in the database.