Improve performance of queries that select GUID columns.
This is based on an idea suggested by @jdanyow. Instead of converting individual bytes of the GUID to strings using `Number#toString` and uppercasing them later, we use a predefined map that allows us to lookup the uppercased hex representation of the individual bytes extremely fast. This easily improves the performance of the GUID string generation by 2500 times.

This fixes #429.