Add dependency on lodash.clonedeep and use that for deep copy.
Existing code is depending on lodash being brought by some other code that
has a dependency on it. That turns out to be not reliable as on older
versions of node, lodash is being placed in the sub-directory of the
dependency that's bringing it in rather than at the top level
node_modules.