Do not remove handlers in 'connect' handler.
Current code does not remove handlers in 'connect' handler for
the socket over which we're going to communicate in
ParallelConnectionStrategy. But 'connect' handlers in other
places do so. The right behavior is to not remove the handlers,
especially the 'error' handler as errors can happen after a
successful connection is established. With the 'connect' handler
we could go either way.