Bug fix and cleanup.
Bug fix: When clientResponse length from getNextBlob() is 0, that
indicates server is done with authentication and start sending data.
The old code was switching to SENT_NTLM_RESPONSE asynchronously,
this was causing problems when the server data comes too soon. Fix is
to switch to SENT_NTLM_RESPONSE synchronously. This is safe as server
should not be sending data unless it's done with authentication on
it's end.

There is an extra round trip involved in authentication process with
NTLM vs Kerberos.