Drop support for node versions that fell out of LTS maintenance window.
Status of various versions of node may be seen at:
https://github.com/nodejs/LTS

Per that, versions v0.10, v0.12 and v5 fell out of maintenance mode. It
might make sense drop support for these versions for future releases of
Tedious.

Specific motivation for this is the Windows Integrated Authentication
work which has test failures due to bugs in those versions, and forcing a
hacky work-around that requires installing latest version of npm before
running the tests for each version of nodejs. See:
https://github.com/tediousjs/tedious/pull/486/commits/15a52b798f14ce90f2a60c2e0000b14b99d9fd97

This part of this PR: https://github.com/tediousjs/tedious/pull/486