Fixed a bug in datefirst validation.
This creeped in PR https://github.com/tediousjs/tedious/pull/503.
Curiously enough the unit tests added as part of the CL caught the bug
when I ran tests on my machine whereas they succeeded on AppVeyor. The
difference is that I happened to have a 'port' config value set in my
local setup whereas AppVeyor did not. Just a freak coincidence that this
would trigger the bug.