chore: Always transpile all code when running unit tests.
When we initially migrated back to ES6+, i added a check here to make sure
that no future JavaScript syntax was being used that was not also supported
by the current Node LTS (4.x at the time).

After having this around for some while, I no longer believe this is worth
the trouble. We should just support all the syntax that was standardized
and is supported by transpilation via babel.