remove scripts/build.js, add build npm script
with the removal of the coffee-script dependency, the scripts/build.js
file ended up just removing the lib dir and spawning the babel cli. It's
much simpler now to have those commands directly in package.json

also, since the "prepublish" event behavior is changing as of npm@5
let's explicitly use the "build" npm event to compile to the lib folder.
the npm "prepublish" event can then be updated to invoke "npm run build"