If encrypt is not specified in the client options but the server requires it, escalate by default.
Changes the behaviour of the pre-login to automatically try again if the server requires
encryption to be enabled but the user hasn't configured the client in that way.

Still allows for the user to explicitly state that they do not ever want to use
encryption but specifying the value as false.

Partially undoes the work done in ceaa8dc15994fe600b01c5e0d84949f3be0984a0 and is a
change in the behaviour, but brings this driver more inline with other TDS
drivers such as jTDS.