fix: Fix deployments to Azure.
The `sspi-client` that was added to support native windows
authentication requires a small amount of C++ code to be compiled via
`node-gyp` during installation. This is not support on Azure out-of-the
box.

As the native windows authentication was never documented and only
accidentally included in the 2.1.0 release, we can remove the code to
support it to fix deployments for Azure users.

Native windows authentication will be brought back and supported in
future `tedious` releases.