chore: Upgrade Node.js versions used for CI.
This updates the Node.js versions that we use to run our tests against during CI on Travis and Appveyor. Node.js 7.x is no longer maintained or supported, 8.x became the new long-term support release, and new development happens on 9.x now.

It also changes the build configuration for Travis to no longer update the installed version of `npm` on every build. This was neccessary when npm introduced the `^` semver range specifier, but all supported versions of Node.js ship with a npm version that supports this out of the box.