feat: add a new `config.authentication` option
This new `config.autehntication` option allows specifying the 
authentication type to be used (`default` or `ntlm`, more to come soon) 
and takes a set of `options` for the specified type.

Existing `config.userName`, `config.password` and `config.domain` 
options will continue working for the time being, but will be removed in 
the future.