fix: fix compatibility with Node.js 11.x
One of the breaking changes in Node.js 11.x is that 
`tls.createSecureContext` will now throw Errors when invalid options are 
given.

In our test cases, we were testing the case where no TLS cipher could be 
negotiated by specifing a cipher suite string of `!ALL`. In Node.js 11, 
this will throw an Error, while in earlier versions this would emit an 
`error` event on the `TLSSocket` instead.

To handle this behaviour, the `tls.createSecureContext` call was moved 
to make sure errors from this call are thrown on `Connection` creation 
where other configuration options are checked for correctness.