fix: change date calculation to be independent of timezones and offsets
This fixes the value calculations for `date`, `datetime`, `datetime2` 
and `datetimeoffset` to be independent of offset changes that could 
cause unexpected values to be written to the database.

BREAKING CHANGE: This should cause the values to be written to the 
database to be more in line of what is expected, but can be different 
for some edge cases.