refactor: use `Readable` streams to generate request payloads
Up until now, request payloads were generated by calling a `.getData` 
method on the request type's payload object, which would call the given 
callback with a `Buffer` containing the full request data.

As this requires the full request data to be fully build (and held in 
memory), actually writing the request to the network is delayed until 
the final `Buffer` was built.

This changeset switches us to use `Readable` stream objects instead. 
Internally, these are still building the full data buffer before, but 
this will be changed in the future.