refactor: generate request payloads using generator functions
When parameters or bulk load rows are converted from JavaScript values to actual TDS request payload data, a method called `writeParameterData` is currently used.

This method takes a `WritableTrackingBuffer` argument and stores the converted data type information in this tracking buffer. As part of https://github.com/tediousjs/tedious/issues/1038, we want to move to a model where this parameter data is not fully stored in one `WritableTrackingBuffer` before a request is written to the network, but instead we want to generate data in a streaming fashion and start sending data over the network as soon as possible.

This changeset adds a new `generate` generator method to all data types, which generates individual `Buffer`s containing the same data as has been written by `WritableTrackingBuffer` so far. `writeParameterData` methods were update to make use of the generator objects returned by `generate`, until all of the request payload generation functionality gets updated to make use of the generator objects directly.