feat: improve performance for large requests
The code that handles the generation of request payloads was rewritten
to make use of JavaScript generators to build payload data in a
just-in-time fashion, instead of generating the whole payload upfront.
This new approach to payload generation greatly reduces memory usage and
the number of garbage collection runs, and indirectly also improves
performance.