fix: ignore connection errors on closed connections
This fixes an issue that could happen when a connection was opened to a non-responding server, and `.close` was called on that connection (e.g. because of the connection timeout being reached).

We would start opening a socket, and when the socket connection failed (because the server could not be reached), the error from the socket would get emitted via the `error` event, which would cause an unhandled exception error if the user did not register an `error` event handler on the connection.

This is fixed by ignoring any errors coming from opening a socket when the connection is already in a closed state.