feat: change default value of `trustServerCertificate`
BREAKING CHANGE: This changes the default value from `true` to `false`. This might lead to encrypted connections not being established if the server's certificate can not be validated against the default list of CAs.

You can either explicitly set `trustServerCertificate` to `true` (not recommended, as it disables certificate validation) or specify a custom CA list via the `cryptoCredentialsDetails` connection option.