chore: add `npm@6.14.13` as a dev dependency
This is mainly to make builds on Node.js 6.x work correctly. The latest version of `semantic-release` pulls in `npm@7.x.x` as a dependency, which is not compatible with Node.js 6.x. We don't really call `semantic-release` on Node.js 6.x, but this dependency causes `npm@7.x.x` to be used when calling `npm` with Node.js 6.x, which causes all `npm` command executions to break.