feat: drop dependency on `readable-stream`
The `readable-stream` module hasn't been updated for a long time, and it looks like it might not receive any updates in the future.

`readable-stream` was based on streams as defined in Node.js 10.x. As this is the oldest version of Node.js we support anyway, including `readable-stream` does not bring us anything. Newer Node.js versions ship with improved but fully backwards-compatible core streams, so we can just use them instead.