feat: remove default value for `clientId` option of `azure-active-directory-password` authentication
BREAKING CHANGE: Using the `azure-active-directory-password`
authentication type now requires specifying the `clientId` option of an
Azure Application. The application has to have public client flows
enabled and has to have `user_impersonation` permissions for Azure SQL
Database. The user account that is being used also has to be assigned
to the application, and has to grant consent (either explicitly or via
admin consent).