Migrate to Flat Config
Change the packages to TypeScript Lint to the next gen package.
Migrate all the settings over from the existing ESLint config.
Remove `deprecated` and `removed` rules.
Enable the typescript strict type checked and stylistic checked rule collections.
Update lint config to ES 2021 instead of ES6.
Add the new ESLint config to the typescript config so that it is aware if it and can provide typing services in the config file.