Fix CLI Args
The next gen version of ES Lint no longer supports passing multiple values with a single instance of the parameter, now it needs to be split into multiple calls of the param.

The ext param is no longer needed however as ESlint knows what it should and should not lint.

Split out the TSC command into a post script since it is technically its own invocation.