ci: improve github actions (#1573)
* ci: use built-in npm caching

The setup-node action automatically stores caches for NPM and uses the
same logic as the current action implementation, so there's no need to
duplicate it in the action definition.

* ci: remove npm upgrade step

The npm upgrade steps in CI only run if the matrix version for node is 6.x
but 6.x is not a tested version of node in any of the matrix definitions.

* ci: only checkout latest history for lint step

The linting step does not need the entire repository history to
run. Only checkout the minimal history to allow faster lint step.

* ci: use setup-sqlserver action for installing SQL Server